Gravitacija.java

+++ Gravitacija.java
@@ -4,15 +4,17 @@ import java.lang.Math;
 public class Gravitacija {
     public static void main(String[] args) {
     	Scanner sc = new Scanner(System.in);
-	int v = sc.nextInt();  //nadmorska visina
+	    double v = sc.nextDouble();  //nadmorska visina
+
         izpisi(v);
-        
         System.out.println("OIS je zakon!");
     }
-    public static void izpisi(int v) {
+
+    public static void izpisi(double v) {
         System.out.println(izracunaj(v));
     }
-    public static double izracunaj(int v) {
+
+    public static double izracunaj(double v) {
 	double C = 6.674 * Math.pow(10, -11); //gravitacijska konstanta
 	double M = 5.972 * Math.pow(10, 24); // masa Zemlje
 	double R = 6.371 * Math.pow(10, 6); //polmer Zemlje
